(fix) squashes bugs in MessageThread.swift and MessageThreadController.swift
in MessageThread.swift fixes bug where container was attempting to be decoded as an array when the proper implementation is a dictionary. This is then mapped in-line to an array to get the desired data structure.
in MessageThreadController.swift - fixes bug in fetchMessageThreads() where we also need a dictionary mapped to an array for messageThreads to be the proper structure. Fixes bug where .resume() was not called after createMessageThreads()'s URLSession.