(test) implements UI tests on MessageThreadsDVC and MessageThreadsTVC.
Test Methods:
testCreateNewMessageThread() - tests for successful completion of creating a new Message Thread.
testTappingOnMessageThreadCell() - tests that cell exists and if it can be tapped on successfully.
testAddNewMessageToThread() - tests for presence of all UI elements and the successful completion of steps neccessary to create and send a new message in a thread.